+++ src/components/create-appointment-fourth/CreateAppointmentFourth.tsx
@@ -11,7 +11,7 @@ import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import MonetizationOn from 'src/assets/icons/MonetizationOn';
 import RightAction from 'src/assets/icons/RightAction';
-import CreateAppointmentFourthDrawer from 'src/components/create-appointment-fourth/caregiver-drawer/CreateAppointmentFourthDrawer';
+import CreateAppointmentFourthDrawer from 'src/components/create-appointment-fourth/drawer/CreateAppointmentFourthDrawer';
 import { SMALL_CAREGIVER_AVATAR_SIZE } from 'src/components/create-appointment-fourth/constants';
 import {
   getMockCaregiverAvatar,
